Calculate the parameter trend

Calculate the parameter trend

Usage

calc_par_trend(
  val_info,
  val_line = c("mean", "sd", "cv", "ar_1_phi"),
  ref_parm_value,
  vals_df,
  parname,
  parlist,
  ctl,
  par_section,
  dat
)

Arguments

val_info

The line in the input df containing info about the parameter.

val_line

Which line in val info to use.

ref_parm_value

This is the historic parameter that the end trend value. Can be NA if the there is no line in val_info for the given parameter

vals_df

The dataframe of the parameter values by year. Use to get start val and last year

parname

Name of the parameter with devs from the SS model. will reference, if using a relative method.

parlist

A parameter file as read in by r4ss::SS_readpar_3.30

ctl

A control file as read in by r4ss::SS_readctl.

par_section

Which parameter section should this variabile be in?

dat

A data file as read in by r4ss::SS_readdat.

Value

A vector of values with length ncol(vals_df), the number of future years.

Author

Kathryn Doering