Update a sequence of base parameter annual values to account for a time varying block effects

Update a sequence of base parameter annual values to account for a time varying block effects

Usage

update_basevals_blocks(
  base_vals,
  base_years,
  temp_block,
  current_par,
  ctl,
  dat,
  temp_ctl,
  base_range,
  baseparm,
  base_bounds
)

Arguments

base_vals

A vector of base parameter values that will be updated to include the impact of a time varying block change

base_years

A vector of years for which the base values are needed

temp_block

The timevarying parameter lines for the block effects on the base parameter

current_par

The index of the current parameter being updated

ctl

A control file as read in by r4ss::SS_readctl.

dat

A data file as read in by r4ss::SS_readdat.

temp_ctl

A subset of the control file representing the parameter section of interest (i.e. MG, SR, Q, or Selectivity)

base_range

the difference between the base parameters max and min bounds

baseparm

The value of the base parameter

base_bounds

The min and max bounds of the base parameter

Value

A modified parameter value series that incorporates the appropriate time varying block effects.

Author

Nathan Vaughan